+++ CHANGELOG.md
@@ -7,13 +7,20 @@ Please enumerate all user-facing changes using format `<githib issue/pr number>:
 
 ## 0.7.0
 
+* [#143](https://github.com/kroxylicious/kroxylicious-junit5-extension/issues/143): Default native kafka image tag from the version of Kafka on the classpath.
 * [#197](https://github.com/kroxylicious/kroxylicious-junit5-extension/issues/197): Consumer of the extension should be able to choose the version of the kafka dependency.
+
 
 ### Changes, deprecations and removals
 
-* It is now necessary to declare org.apache.kafka:kafka_2.13 as a test dependency.  If you wish to use kafka in-vm with zookeeper, org.apache.zookeeper:zookeeper
-  must be a test dependency too.
-
+* It is now necessary to declare `org.apache.kafka:kafka_2.13` as a test dependency.
+* If you wish to use kafka in-vm with zookeeper, `org.apache.zookeeper:zookeeper` must be declared as test dependency too.
+* The way `TestcontainersKafkaCluster` determines the native image version tag to use for kafka and zookeeper is changed.
+  Previously `TestcontainersKafkaCluster` always used the `latest-snapshot` version.  With this release, it will now default
+  to use the version of the kafka native image that corresponds to the version of Kafka Broker found on the classpath.
+  This will give consumers of the test extension that have tests using `TestcontainersKafkaCluster` test repeatability
+  and consumers who use both `InVMKafkaCluster` and `TestcontainersKafkaCluster` kafka broker version consistency.
+  To get back the original behaviour, annotate the  `KafkaCluster` with `@Version("latest-snapshot")`.
 
 ## 0.6.0